Jpa repository, java 11, pmd
- Used JpaRepository for database interaction instead of entityManager. Refactored the code for the same.
- Moved to Java 11 to use more of its features.
- Added pmd configuration
