+++ pom.xml
@@ -21,7 +21,9 @@
     <properties>
         <project.build.sourceEncoding>UTF-8</project.build.sourceEncoding>
         <project.reporting.outputEncoding>UTF-8</project.reporting.outputEncoding>
-        <java.version>1.8</java.version>
+        <java.version>11</java.version>
+        <start-class>com.simplified.aanchal.Application</start-class>
+        <pmd-config-url>${project.basedir}/mvnconf/etc/aanchal_pmd.xml</pmd-config-url>
     </properties>
 
     <dependencies>
