+++ src/main/java/com/simplified/aanchal/connectors/db/CustomerDAO.java
@@ -1,0 +1,16 @@
+package com.simplified.aanchal.connectors.db;
+
+import java.util.List;
+
+public interface CustomerDAO {
+	
+	List<Customer> findAll();
+	
+	Customer findById(int id);
+	
+	void save(Customer customer);
+	
+	void deleteById(int id);
+	
+
+}
