--- src/main/java/com/simplified/aanchal/dao/CustomerDAO.java
@@ -1,18 +1,0 @@
-package com.simplified.aanchal.dao;
-
-import java.util.List;
-
-import com.simplified.aanchal.modal.Customer;
-
-public interface CustomerDAO {
-	
-	List<Customer> get();
-	
-	Customer get(int id);
-	
-	void save(Customer customer);
-	
-	void delete(int id);
-	
-
-}
