+++ src/main/resources/application.properties
@@ -11,3 +11,6 @@ server.port=9780
 # By default, the master change log is read from db/changelog/db.changelog-master.yaml,
 # but you can change the location by setting spring.liquibase.change-log.
 spring.liquibase.change-log=classpath:db/changelog/db.changelog-master.xml
+
+#log level
+logging.level.root=INFO
